CRM-19787 Filter available contacts by relationship sub-type when adding assigning a role in CiviCase
Without this patch, when assigning a CiviCase role where the relationship
is defined to apply to a sub-type, all contacts are shown. If a contact
of the wrong sub-type is selected, the assignment fails.

This patch filters the displayed contacts to those of the specified
sub-type.

This includes the fix in https://github.com/civicrm/civicrm-core/pull/9560